Back in the Game of Coding
Shifts ownership to LittleRover, changes all the licenses to GNU GPLv3.

Applies stricter formatting rules and warning awareness, although plenty
of warnings are not yet taken care of.

Camera is being prepped for a smarter implementation, having been
stripped of currently unused methods and fields.

Adds an unfinished frustrum object.

Realises that the current creature package implementations are severely
incomplete and need a proper visit.

Exceptions are revised to contain only relevant constructors. Exception
loggers need to be reconsidered carefully. AuroraLogs class is now used
in every class that has a logger and will be the logger used in every
class.

Added IOUtils for reading certain files in single lines.

Main is acknowledged as terrible, will instead have a main for Client
and a main for Server which will be where the program is launched with
limited command line functions in future.

Added a math.geometry package which will include math involving volumes,
planes, lines, and points. Currently contains unfinished implementations
of volumes and planes.

LD and HD versions of Vec, Ang, and Ref have been revised to be more
flexible. Notably this makes vectors larger as they are now quaternions
however, the transition of Ref being only affine matricies to being
general 4x4 matricies should make them more flexible. Ref is now
refactored to be Mat. HDQ implementations still need to be done but are
lower priority.

Revisited the model system and begun implementing a model object that
should work with shaders in a flexible fashion. As this will be a
bottleneck this is likely to be reconstructed multiple times in future.
Plan is to have a separate renderer which models are passed through so
different GL versions can be made use of depending on hardware.
GLMaterial is being reconsidered as an implementation of a shader and
collection of textures, but has not been implemented yet. Added seperate
vertex and face objects for future model generation code, especially
when it comes to custom creatures and custom terrain.

Implements as GLSharedObject as a special case of GLObject that is
specialised for being used by multiple GLObjects, like materials, model
buffers, textures, and things like that. Note: GLTexture hasn't yet had
this applied.

Altered the GLVersion tag to use a number instead of string for version.
Still need to make it useful in some way, but it may be something to
instead discard as effectively useless.

Shaders is reconstructed again, but is unfinished.

Shift away from the idea of a Player Soul which will limit future
implementations to instead just a player object that will apply much
more generally.

Note that the current statistical implementations are very rigid and
will not survive future changes.

Adds some old ideas for the greater scale of the world, although these
are overly ambitious and likely to get revised and cut or maintained in
a mostly unused form.